requirements.txt
## The following requirements were added by pip freeze:
boto3==1.16.35
botocore==1.19.35
jmespath==0.10.0
python-dateutil==2.8.1
PyYAML==5.3.1
s3transfer==0.3.3
Send2Trash==1.5.0
six==1.15.0
urllib3==1.26.2
